add customer model and use bigDecimal replace double for price and amount

+++ domain/src/main/java/com/thoughtworks/dddpractice/domain/goods/exception/PriceLessThanZeroException.java
@@ -2,9 +2,11 @@ package com.thoughtworks.dddpractice.domain.goods.exception;
 
 import com.thoughtworks.dddpractice.framework.annotations.domain.DomainException;
 
+import java.math.BigDecimal;
+
 @DomainException
 public class PriceLessThanZeroException extends RuntimeException {
-  public PriceLessThanZeroException(Double price) {
+  public PriceLessThanZeroException(BigDecimal price) {
     super("Goods price " + price + " is less than 0");
   }
 }
